Move abeja.models.api.APIClient to abeja.training
